+++ seisflows/optimize/lib/LBFGS.py
@@ -3,7 +3,7 @@ import numpy as np
 
 from seisflows.tools import unix
 from seisflows.tools.array import loadnpy, savenpy
-from seisflows.tools.code import savetxt, exists
+from seisflows.tools.tools import savetxt, exists
 from seisflows.tools.math import angle
