+++ seisflows/optimize/lib/NLCG.py
@@ -6,7 +6,7 @@ from seisflows.tools import unix
 
 from seisflows.tools.array import loadnpy, savenpy
 from seisflows.tools.math import dot
-from seisflows.tools.code import loadtxt, savetxt
+from seisflows.tools.tools import loadtxt, savetxt
