+++ seisflows/optimize/lib/PLCG.py
@@ -4,7 +4,7 @@ from os.path import join
 import numpy as np
 
 from seisflows.tools import unix
-from seisflows.tools.code import loadtxt, savetxt
+from seisflows.tools.tools import loadtxt, savetxt
 #from seisflows.tools.io import OutputWriter
 
 from seisflows.optimize.lib.LBFGS import LBFGS
