+++ seisflows/postprocess/base.py
@@ -5,7 +5,7 @@ import numpy as np
 from os.path import join
 from seisflows.tools import unix
 from seisflows.tools.array import savenpy
-from seisflows.tools.code import exists
+from seisflows.tools.tools import exists
 from seisflows.config import ParameterError
 
 PAR = sys.modules['seisflows_parameters']
