+++ seisflows/solver/specfem2d.py
@@ -10,7 +10,7 @@ from seisflows.tools.shared import getpar, setpar
 
 from seisflows.tools import unix
 from seisflows.tools.array import loadnpy, savenpy
-from seisflows.tools.code import exists, call_solver, call_solver_nompi
+from seisflows.tools.tools import exists, call_solver, call_solver_nompi
 from seisflows.config import   \
     ParameterError, custom_import
