+++ seisflows/system/multithreaded.py
@@ -1,17 +1,15 @@
 
 import os
 import sys
+import numpy as np
 
 from os.path import abspath, basename, join
 from subprocess import Popen
 from time import sleep
 
-import numpy as np
-
 from seisflows.tools import unix
-from seisflows.tools.code import call, findpath, saveobj
-from seisflows.config import   \
-    ParameterError, custom_import
+from seisflows.tools.tools import call, findpath, saveobj
+from seisflows.config import ParameterError, custom_import
 
 PAR = sys.modules['seisflows_parameters']
 PATH = sys.modules['seisflows_paths']
