+++ seisflows/system/slurm_sm.py
@@ -4,7 +4,7 @@ import sys
 
 from os.path import abspath, basename, join
 from seisflows.tools import unix
-from seisflows.tools.code import call, findpath, saveobj
+from seisflows.tools.tools import call, findpath, saveobj
 from seisflows.config import ParameterError, custom_import
 
 PAR = sys.modules['seisflows_parameters']
