+++ seisflows/workflow/inversion.py
@@ -6,7 +6,7 @@ import numpy as np
 from seisflows.tools import msg
 from seisflows.tools import unix
 from seisflows.tools.array import loadnpy, savenpy
-from seisflows.tools.code import divides, exists
+from seisflows.tools.tools import divides, exists
 from seisflows.config import ParameterError
 
 PAR = sys.modules['seisflows_parameters']
