+++ seisflows/workflow/migration.py
@@ -3,7 +3,7 @@ import sys
 from glob import glob
 
 from seisflows.tools import unix
-from seisflows.tools.code import exists
+from seisflows.tools.tools import exists
 from seisflows.config import ParameterError
 
 PAR = sys.modules['seisflows_parameters']
