+++ seisflows/workflow/test_adjoint.py
@@ -5,7 +5,7 @@ import numpy as np
 from glob import glob
 from os.path import basename, join
 from seisflows.tools import unix
-from seisflows.tools.code import exists
+from seisflows.tools.tools import exists
 from seisflows.config import ParameterError
