+++ seisflows/workflow/test_postprocess.py
@@ -4,7 +4,7 @@ import numpy as np
 
 from seisflows.tools import unix
 from seisflows.tools.array import loadnpy, savenpy
-from seisflows.tools.code import exists
+from seisflows.tools.tools import exists
 from seisflows.config import custom_import, ParameterError
 
 PAR = sys.modules['seisflows_parameters']
